+++ src/main/kotlin/io/snyk/plugin/ui/SnykSettingsDialog.kt
@@ -181,7 +181,7 @@ class SnykSettingsDialog(
             // TODO: check for concrete project roots, and if we received a message for them
             // this is an edge case, when a project is opened after ls initialization and
             // preferences dialog is opened before ls sends the additional parameters
-            additionalParametersTextField.isEnabled = LanguageServerWrapper.getInstance().folderConfigsRefreshed.isNotEmpty()
+            additionalParametersTextField.isEnabled = LanguageServerWrapper.getInstance(project).getFolderConfigsRefreshed().isNotEmpty()
             additionalParametersTextField.text = getAdditionalParams(project)
             scanOnSaveCheckbox.isSelected = applicationSettings.scanOnSave
             cliReleaseChannelDropDown.selectedItem = applicationSettings.cliReleaseChannel
@@ -458,7 +458,7 @@ class SnykSettingsDialog(
 
         runAsync {
             if (!pluginSettings().isGlobalIgnoresFeatureEnabled) {
-                LanguageServerWrapper.getInstance().refreshFeatureFlags()
+                LanguageServerWrapper.getInstance(project).refreshFeatureFlags()
             }
             issueViewPanel.isVisible = pluginSettings().isGlobalIgnoresFeatureEnabled
         }